feature(onboarding): refactoring call friends view
